card_service-main/requirements.txt
aiohttp==3.7.3
grpcio
grpcio-tools
aioprometheus[aiohttp,binary]
aioredis==1.3.1
